--- tests/fixtures/auth.json
@@ -1,18 +1,0 @@
-{
-  "cookies": [],
-  "origins": [
-    {
-      "origin": "http://localhost:5173",
-      "localStorage": [
-        {
-          "name": "sb-qdspccxbmpaozsufbzuf-auth-token",
-          "value": "{\"access_token\":\"eyJhbGciOiJIUzI1NiIsImtpZCI6ImNMVGRJMHhuTzB1ZHRZWUgiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL3Fkc3BjY3hibXBhb3pzdWZienVmLnN1cGFiYXNlLmNvL2F1dGgvdjEiLCJzdWIiOiI1NGVlZDE4Zi00MTA3LTRiZDgtOTI4Ny03OGQ1OGQ5YWY4OGMiLCJhdWQiOiJhdXRoZW50aWNhdGVkIiwiZXhwIjoxNzQ0NzgzMDA3LCJpYXQiOjE3NDQ3Nzk0MDcsImVtYWlsIjoiZWhqYW5nQGdtYWlsLmNvbSIsInBob25lIjoiIiwiYXBwX21ldGFkYXRhIjp7InByb3ZpZGVyIjoiZW1haWwiLCJwcm92aWRlcnMiOlsiZW1haWwiXX0sInVzZXJfbWV0YWRhdGEiOnsiZW1haWwiOiJlaGphbmdAZ21haWwuY29tIiwiZW1haWxfdmVyaWZpZWQiOnRydWUsIm5pY2tuYW1lIjoi7Jy864SkIiwicGhvbmVfdmVyaWZpZWQiOmZhbHNlLCJwcm9maWxlX2ltYWdlIjoiZGF0YTppbWFnZS9zdmcreG1sLCUzY3N2ZyUyMHdpZHRoPScxMDAnJTIwaGVpZ2h0PScxMDAnJTIwdmlld0JveD0nMCUyMDAlMjAxMDAlMjAxMDAnJTIwZmlsbD0nbm9uZSclMjB4bWxucz0naHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmcnJTNlJTNjZyUyMGNsaXAtcGF0aD0ndXJsKCUyM2NsaXAwXzU2NV80ODcpJyUzZSUzY3JlY3QlMjB3aWR0aD0nMTAwJyUyMGhlaWdodD0nMTAwJyUyMGZpbGw9JyUyMzNBM0QzRCcvJTNlJTNjcGF0aCUyMGQ9J001MC42NjY3JTIwNjIuMzMzNEM2MC4wNTU2JTIwNjIuMzMzNCUyMDY3LjY2NjclMjA1NC43MjIyJTIwNjcuNjY2NyUyMDQ1LjMzMzRDNjcuNjY2NyUyMDM1Ljk0NDUlMjA2MC4wNTU2JTIwMjguMzMzNCUyMDUwLjY2NjclMjAyOC4zMzM0QzQxLjI3NzklMjAyOC4zMzM0JTIwMzMuNjY2NyUyMDM1Ljk0NDUlMjAzMy42NjY3JTIwNDUuMzMzNEMzMy42NjY3JTIwNTQuNzIyMiUyMDQxLjI3NzklMjA2Mi4zMzM0JTIwNTAuNjY2NyUyMDYyLjMzMzRaJyUyMGZpbGw9JyUyM0IwQjBCMCcvJTNlJTNjcGF0aCUyMGQ9J001MC42NjY4JTIwNjUuNjY2N0MzNC42NTc5JTIwNjUuNjg3OCUyMDIxLjY4NDYlMjA4MC45ODczJTIwMjEuNjY2NyUyMDk5Ljg2NjhDMjEuNjY2NyUyMDEwMS45NjUlMjAyMy4xMDk0JTIwMTAzLjY2NyUyMDI0Ljg4ODklMjAxMDMuNjY3SDc2LjQ0NDZDNzguMjI0MSUyMDEwMy42NjclMjA3OS42NjY3JTIwMTAxLjk2NSUyMDc5LjY2NjclMjA5OS44NjY4Qzc5LjY0OTElMjA4MC45ODczJTIwNjYuNjc1OCUyMDY1LjY4NzYlMjA1MC42NjY4JTIwNjUuNjY2N1onJTIwZmlsbD0nJTIzQjBCMEIwJy8lM2UlM2MvZyUzZSUzY2RlZnMlM2UlM2NjbGlwUGF0aCUyMGlkPSdjbGlwMF81NjVfNDg3JyUzZSUzY3JlY3QlMjB3aWR0aD0nMTAwJyUyMGhlaWdodD0nMTAwJyUyMHJ4PSc1MCclMjBmaWxsPSd3aGl0ZScvJTNlJTNjL2NsaXBQYXRoJTNlJTNjL2RlZnMlM2UlM2Mvc3ZnJTNlIiwic3ViIjoiNTRlZWQxOGYtNDEwNy00YmQ4LTkyODctNzhkNThkOWFmODhjIn0sInJvbGUiOiJhdXRoZW50aWNhdGVkIiwiYWFsIjoiYWFsMSIsImFtciI6W3sibWV0aG9kIjoicGFzc3dvcmQiLCJ0aW1lc3RhbXAiOjE3NDQ3Nzk0MDd9XSwic2Vzc2lvbl9pZCI6IjA1N2UzMmM4LWJlNmItNGY1Mi04NzVkLTY1OGJlOGYyMTFkYiIsImlzX2Fub255bW91cyI6ZmFsc2V9.0RgxJYyUbjx63ne8dwqRtNxpRHd4QLYNz2HvFE-lniM\",\"token_type\":\"bearer\",\"expires_in\":3600,\"expires_at\":1744783007,\"refresh_token\":\"-yOE5Pqswkn8bYTE9hF59w\",\"user\":{\"id\":\"54eed18f-4107-4bd8-9287-78d58d9af88c\",\"aud\":\"authenticated\",\"role\":\"authenticated\",\"email\":\"ehjang@gmail.com\",\"email_confirmed_at\":\"2025-04-15T02:46:49.483114Z\",\"phone\":\"\",\"confirmed_at\":\"2025-04-15T02:46:49.483114Z\",\"last_sign_in_at\":\"2025-04-16T04:56:47.958625394Z\",\"app_metadata\":{\"provider\":\"email\",\"providers\":[\"email\"]},\"user_metadata\":{\"email\":\"ehjang@gmail.com\",\"email_verified\":true,\"nickname\":\"으네\",\"phone_verified\":false,\"profile_image\":\"data:image/svg+xml,%3csvg%20width='100'%20height='100'%20viewBox='0%200%20100%20100'%20fill='none'%20xmlns='http://www.w3.org/2000/svg'%3e%3cg%20clip-path='url(%23clip0_565_487)'%3e%3crect%20width='100'%20height='100'%20fill='%233A3D3D'/%3e%3cpath%20d='M50.6667%2062.3334C60.0556%2062.3334%2067.6667%2054.7222%2067.6667%2045.3334C67.6667%2035.9445%2060.0556%2028.3334%2050.6667%2028.3334C41.2779%2028.3334%2033.6667%2035.9445%2033.6667%2045.3334C33.6667%2054.7222%2041.2779%2062.3334%2050.6667%2062.3334Z'%20fill='%23B0B0B0'/%3e%3cpath%20d='M50.6668%2065.6667C34.6579%2065.6878%2021.6846%2080.9873%2021.6667%2099.8668C21.6667%20101.965%2023.1094%20103.667%2024.8889%20103.667H76.4446C78.2241%20103.667%2079.6667%20101.965%2079.6667%2099.8668C79.6491%2080.9873%2066.6758%2065.6876%2050.6668%2065.6667Z'%20fill='%23B0B0B0'/%3e%3c/g%3e%3cdefs%3e%3cclipPath%20id='clip0_565_487'%3e%3crect%20width='100'%20height='100'%20rx='50'%20fill='white'/%3e%3c/clipPath%3e%3c/defs%3e%3c/svg%3e\",\"sub\":\"54eed18f-4107-4bd8-9287-78d58d9af88c\"},\"identities\":[{\"identity_id\":\"0c641bb9-c8c0-48b4-802d-b44f9b63d457\",\"id\":\"54eed18f-4107-4bd8-9287-78d58d9af88c\",\"user_id\":\"54eed18f-4107-4bd8-9287-78d58d9af88c\",\"identity_data\":{\"email\":\"ehjang@gmail.com\",\"email_verified\":false,\"nickname\":\"으네\",\"phone_verified\":false,\"profile_image\":\"data:image/svg+xml,%3csvg%20width='100'%20height='100'%20viewBox='0%200%20100%20100'%20fill='none'%20xmlns='http://www.w3.org/2000/svg'%3e%3cg%20clip-path='url(%23clip0_565_487)'%3e%3crect%20width='100'%20height='100'%20fill='%233A3D3D'/%3e%3cpath%20d='M50.6667%2062.3334C60.0556%2062.3334%2067.6667%2054.7222%2067.6667%2045.3334C67.6667%2035.9445%2060.0556%2028.3334%2050.6667%2028.3334C41.2779%2028.3334%2033.6667%2035.9445%2033.6667%2045.3334C33.6667%2054.7222%2041.2779%2062.3334%2050.6667%2062.3334Z'%20fill='%23B0B0B0'/%3e%3cpath%20d='M50.6668%2065.6667C34.6579%2065.6878%2021.6846%2080.9873%2021.6667%2099.8668C21.6667%20101.965%2023.1094%20103.667%2024.8889%20103.667H76.4446C78.2241%20103.667%2079.6667%20101.965%2079.6667%2099.8668C79.6491%2080.9873%2066.6758%2065.6876%2050.6668%2065.6667Z'%20fill='%23B0B0B0'/%3e%3c/g%3e%3cdefs%3e%3cclipPath%20id='clip0_565_487'%3e%3crect%20width='100'%20height='100'%20rx='50'%20fill='white'/%3e%3c/clipPath%3e%3c/defs%3e%3c/svg%3e\",\"sub\":\"54eed18f-4107-4bd8-9287-78d58d9af88c\"},\"provider\":\"email\",\"last_sign_in_at\":\"2025-04-15T02:46:49.479904Z\",\"created_at\":\"2025-04-15T02:46:49.479949Z\",\"updated_at\":\"2025-04-15T02:46:49.479949Z\",\"email\":\"ehjang@gmail.com\"}],\"created_at\":\"2025-04-15T02:46:49.475629Z\",\"updated_at\":\"2025-04-16T04:56:47.961088Z\",\"is_anonymous\":false}}"
-        },
-        {
-          "name": "user-storage",
-          "value": "{\"state\":{\"user\":{\"id\":\"54eed18f-4107-4bd8-9287-78d58d9af88c\",\"email\":\"ehjang@gmail.com\",\"nickname\":\"으네\",\"profile_image\":\"data:image/svg+xml,%3csvg%20width='100'%20height='100'%20viewBox='0%200%20100%20100'%20fill='none'%20xmlns='http://www.w3.org/2000/svg'%3e%3cg%20clip-path='url(%23clip0_565_487)'%3e%3crect%20width='100'%20height='100'%20fill='%233A3D3D'/%3e%3cpath%20d='M50.6667%2062.3334C60.0556%2062.3334%2067.6667%2054.7222%2067.6667%2045.3334C67.6667%2035.9445%2060.0556%2028.3334%2050.6667%2028.3334C41.2779%2028.3334%2033.6667%2035.9445%2033.6667%2045.3334C33.6667%2054.7222%2041.2779%2062.3334%2050.6667%2062.3334Z'%20fill='%23B0B0B0'/%3e%3cpath%20d='M50.6668%2065.6667C34.6579%2065.6878%2021.6846%2080.9873%2021.6667%2099.8668C21.6667%20101.965%2023.1094%20103.667%2024.8889%20103.667H76.4446C78.2241%20103.667%2079.6667%20101.965%2079.6667%2099.8668C79.6491%2080.9873%2066.6758%2065.6876%2050.6668%2065.6667Z'%20fill='%23B0B0B0'/%3e%3c/g%3e%3cdefs%3e%3cclipPath%20id='clip0_565_487'%3e%3crect%20width='100'%20height='100'%20rx='50'%20fill='white'/%3e%3c/clipPath%3e%3c/defs%3e%3c/svg%3e\",\"description\":\"\",\"subscribe_count\":0}},\"version\":0}"
-        }
-      ]
-    }
-  ]
-}
